// vision_estate/frontend/src/lib/generated/api-types.ts
export interface components {
    schemas: {
// ...
        BrokerBooking: components["schemas"]["Booking"] & {
            slot: components["schemas"]["Slot"] & {
                broker: {
                    name: string;
                };
            };
            property: {
                address: components["schemas"]["Address"];
                sellerContact: components["schemas"]["Contact"];
            };
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}